Struct hartex_rust::system::resource_manager::ResourceManager

pub struct ResourceManager {
    resource_control_blocks: [Option<ResourceControlBlock>; 32],
    top: usize,
    pi_stack: [i32; 32],
    curr: usize,
    system_ceiling: i32,
}

Maintains state of all Resources in the Kernel.

Fields
resource_control_blocks: [Option<ResourceControlBlock>; 32]

An Array of ResourceControlBlock for every resource.

top: usize

Points the top of the pi_stack.

pi_stack: [i32; 32]

This stack is used for locking and unlocking of resources.

curr: usize

It next empty index in the array where the new ResourceControlBlock can be stored.

system_ceiling: i32

Hold the ceiling of the resource with the highest ceiling amongst the currently locked resources.

Methods

impl ResourceManager

pub const fn new() -> Self

Returns a new instance of ResourceManager.

pub fn create(
&mut self,
tasks_mask: BooleanVector
) -> Result<ResourceId, KernelError>

Creates a new resource with the task_mask passed as argument. The new resource created is then pushed into resource_control_blocks.

pub fn lock(&mut self, id: ResourceId, curr_tid: TaskId) -> Option<u32>

This procedure locks the resource from further use. It is freed when unlock is called. By locking a resource, this function first checks if the task with TaskId curr_tid has access to the resource. If not, then return None. But If it has access, then generate the task_mask corresponding to the tasks that have to be blocked and return it.

Note that though the resource is locked, it only blocks the tasks below its ceiling priority and not the others. Hence If any task with higher priority is released, it can lock resources with a higher ceiling. But Since only a task with higher priority can execute, it won’t allow the scheduler to execute tasks that would cause a cyclic wait for resources hence avoiding deadlocks.

fn get_pi_mask(&self, ceiling: TaskId) -> u32

Returns the Pi_mask, which is just a boolean vector with all bits up to ceiling (including) set to 1.

pub fn unlock(&mut self, id: ResourceId) -> Option<u32>

This procedure would unlock the resource for other tasks if it were the lastly locked resource. It calls pop_stack, which pops the top of the stack and also re-assigns the system_ceiling. It returns a BooleanVector corresponding to the tasks that have to be unblocked.

fn pop_stack(&mut self)

Pops the stack top and assigns the system_ceiling to the new stack top.

fn push_stack(&mut self, ceiling: TaskId)

Pushes the passed ceiling onto the pi_stack.