Struct hartex_rust::system::task_manager::Scheduler

#[repr(C)]
pub struct Scheduler {
    pub curr_tid: usize,
    pub is_running: bool,
    pub task_control_blocks: [Option<TaskControlBlock>; 32],
    pub blocked_tasks: BooleanVector,
    pub active_tasks: BooleanVector,
    pub is_preemptive: bool,
    pub started: bool,
}

Maintains state of all tasks in the Kernel

Fields
curr_tid: usize

The Task id of the currently running task.

is_running: bool

True if the scheduler has started scheduling tasks on the CPU.

task_control_blocks: [Option<TaskControlBlock>; 32]

An Array of task control blocks corresponding to each task (created only if task exists).

blocked_tasks: BooleanVector

A boolean vector in which, if a bit at a position is true, it implies that the task is active and to be scheduled.

active_tasks: BooleanVector

A boolean vector in which, if a bit at a position is true, it implies that the task is blocked and cannot be scheduled even if it’s active.

is_preemptive: bool

A variable which decided if the scheduler should preemptively schedule tasks or not.

started: bool

Set true as soon as the first task is scheduled.

Methods

impl Scheduler

pub const fn new() -> Self

Returns a new instance of Scheduler

pub fn init(&mut self, is_preemptive: bool)

This method sets the is_preemptive field of the scheduler instance and defines the configurations for the idle task and calls create_task with it. The waiting task has zero priority; hence, it is only executed when no other task is in Ready state.

pub fn start_kernel(&mut self)

It sets the is_running field to true as kernel then starts scheduling tasks.

pub fn create_task<T: Sized>(
&mut self,
priority: usize,
stack: &mut [u32],
handler_fn: fn(_: &T) -> !,
param: &T
) -> Result<(), KernelError> where
T: Sync,

The program counter for the task is pointer value of the function pointer (handler_fn). param is a variable whose reference will be made accessible to the task, and this helps in sharing global state with other tasks. Both these values are stored in a specific index of the stack so that when the context_switch function loads the stack for this task, the appropriate program counter and argument for that function is loaded. An important thing to note is that the task’s index in the task_control_blocks is the priority of the task. Hence there can be only one task of the mentioned priority. Also, another important thing is that the argument param is of a generic type(T).

The <T: Sync> informs the compiler that the type T must implement the Sync trait. By implementing the Sync trait, a type becomes safe to be shared across tasks. Hence if a type that doesn’t implement Sync trait (like a mutable integer) is passed as param, then the code won’t compile. Kernel primitives like Message and Resource (which are data race safe) implement the Sync trait; hence, it can be passed as param. In this way, the Kernel makes safety a requirement rather than a choice.

handler_fn is of type fn(&T) -> !, which implies it is a function pointer which takes a parameter of Type &T and infinitely loops. For more details, look into spawn! Macro.

fn create_tcb<T: Sized>(
&self,
stack: &mut [u32],
handler: fn(_: &T) -> !,
param: &T
) -> Result<TaskControlBlock, KernelError> where
T: Sync,

Creates a TCB corresponding to the tasks details passed onto this method.

fn insert_tcb(
&mut self,
idx: usize,
tcb: TaskControlBlock
) -> Result<(), KernelError>

Inserts the TCB into task_control_blocks at position idx.

pub fn block_tasks(&mut self, tasks_mask: BooleanVector)

Appends tasks_mask onto blocked_tasks.

pub fn unblock_tasks(&mut self, tasks_mask: BooleanVector)

Removes tasks_mask from blocked_tasks.

pub fn get_next_tid(&self) -> TaskId

Returns the TaskId currently high priority task, which is in ready state. The highest priority is determined by calculating the most significant bit of boolean vector corresponding to the tasks in the ready state. The tasks in the ready state can be identified by the boolean and of active_tasks and boolean not(blocked_tasks).

pub fn release(&mut self, tasks_mask: BooleanVector)

Updates active_tasks with task_mask.