Module hartex_rust::types

Exports types defined across other Kernel modules.

Structs

Message
It holds a variable of a generic type so that any message can be stored in it. It designed to work without ending up in data races.

Resource
A Safe Container to store a resource, it can hold resource of any Generic Type and allow safe access to it without leading into Data races or Deadlocks.

Enums

EventTableType
Events are executed at multiples of certain time units. For example, an event can be dispatched once every 40 milliseconds, or 50 seconds, etc. This enum represents each such time units. Note that this includes OnOff also as OnOff events do not belong to the category of time units; hence, it is given as a separate field.

EventType

Type Definitions

BooleanVector

EventId

MessageId

ResourceId

SemaphoreId

TaskId